requirements.txt
Django==4.2
Flask==3.1.3
python-decouple==3.8
python-dotenv==1.2.2
requests==2.32.5
pillow==12.1.1
pillow_heif==1.3.0
pywebpush==2.3.0
cryptography==46.0.0
gunicorn==21.2.0
whitenoise==6.6.0
aiogram==3.26.0
psycopg2-binary==2.9.11
piexif>=1.1.3  # EXIF обработка